+++ src/mqt/core/_core/symbolic.pyi
@@ -1,9 +1,30 @@
 from collections.abc import Iterator, Mapping, Sequence
 from typing import overload
 
-from ..._compat.typing import Self
-from .term import Term
-from .variable import Variable
+from .._compat.typing import Self
+
+class Variable:
+    def __eq__(self: Self, arg0: object) -> bool: ...
+    def __gt__(self: Self, arg0: Variable) -> bool: ...
+    def __init__(self: Self, name: str = "") -> None: ...
+    def __lt__(self: Self, arg0: Variable) -> bool: ...
+    def __ne__(self: Self, arg0: object) -> bool: ...
+    @property
+    def name(self: Self) -> str: ...
+
+class Term:
+    def __init__(self: Self, variable: Variable, coefficient: float = 1.0) -> None: ...
+    def __mul__(self: Self, arg0: float) -> Term: ...
+    def __rmul__(self: Self, arg0: float) -> Term: ...
+    def __rtruediv__(self: Self, arg0: float) -> Term: ...
+    def __truediv__(self: Self, arg0: float) -> Term: ...
+    def add_coefficient(self: Self, coeff: float) -> None: ...
+    def evaluate(self: Self, assignment: Mapping[Variable, float]) -> float: ...
+    def has_zero_coefficient(self: Self) -> bool: ...
+    @property
+    def coefficient(self: Self) -> float: ...
+    @property
+    def variable(self: Self) -> Variable: ...
 
 class Expression:
     constant: float
