--- src/mqt/core/_core/symbolic/__init__.pyi
@@ -1,9 +1,0 @@
-from .expression import Expression
-from .term import Term
-from .variable import Variable
-
-__all__ = [
-    "Expression",
-    "Term",
-    "Variable",
-]
