--- src/mqt/core/_core/symbolic/variable.pyi
@@ -1,10 +1,0 @@
-from ..._compat.typing import Self
-
-class Variable:
-    def __eq__(self: Self, arg0: object) -> bool: ...
-    def __gt__(self: Self, arg0: Variable) -> bool: ...
-    def __init__(self: Self, name: str = "") -> None: ...
-    def __lt__(self: Self, arg0: Variable) -> bool: ...
-    def __ne__(self: Self, arg0: object) -> bool: ...
-    @property
-    def name(self: Self) -> str: ...
