+++ src/python/module.cpp
@@ -12,8 +12,7 @@ void registerSymbolic(py::module& m);
 void registerQuantumComputation(py::module& m);
 
 PYBIND11_MODULE(_core, m) {
-  py::module permutation = m.def_submodule("permutation");
-  registerPermutation(permutation);
+  registerPermutation(m);
 
   py::module symbolic = m.def_submodule("symbolic");
   registerSymbolic(symbolic);
@@ -21,8 +20,7 @@ PYBIND11_MODULE(_core, m) {
   py::module operations = m.def_submodule("operations");
   registerOperations(operations);
 
-  py::module quantumComputation = m.def_submodule("quantum_computation");
-  registerQuantumComputation(quantumComputation);
+  registerQuantumComputation(m);
 }
 
 } // namespace mqt
